Reorganise files to separate GildedRose and Item classes
This commit separates the `GildedRose` and `Item` classes into separate
files. We are also creating folder to move the relevant files to:

- lib folder to move the `GildedRose` and `Item` classes to,
- spec folder to move all spec files to,
- test folder to move all test files to,
- texttest folder to move the texttest files to.

As we are moving the `Item` class into its own file and are not allowed
to make changes to this class itself, we have also added a spec file for
this class. This should ensure that any upcoming changes we are making
for this application will not break this class.