Extract methods to increase/decrease quality
Another small refactor we can make at this stage is to extract methods
for when we increase or decrease the quality value by one, which makes
it easier to read. For now, we are keeping the decrease of sell in value
as is and move it higher up since this is applied only once and
happening to any item.