Move update logic to delegator class
In an attempt to move out logic of updating the items out of the
`GildedRose` class, we are introducing a delegator class `ItemDelegator`
that can take in an `Item` object and add methods to call on them. As a
result, the `GildedRose` class no longer know about the specifics and
behaviours of the `Item` class.