Define sub classes for special items
We are now setting up individual classes for the special items: "Aged
Brie", "Backstage passes to a TAFKAL80ETC concert", "Sulfuras, Hand of
Ragnaros", which are inheriting from the `ItemDelegator` class.

As a result, the `ItemDelegator` class can deal with updating the
quality and sell in value for all standard items. In case of updating
a special item, it knows how to delegate to the specific sub classes.